Change database from H2 to PostgreSQL
Changes database to PostgresSQL. Defines ddl-auto with CREATE value to recreate database every time the application is restarted because at this point of the project a lot of changes will be done on database and there is not production environment to maintain. Also defines show-sql with TRUE value to follow how Hibernate is generation the tables.