Improve error handling to return standardized error responses
The error responses will have an error id so the specific error of the request can be easily be found on logs. Removed generic error handling that was causing wrong http status to be returned (it was always returning 500) and move the logic to populate the error response fields to CustomErrorAttributes.