Add integration and controller tests for authentication feature
The integration tests are just to test the happy flow of signup and
login, the controller tests also tests how the API behaves in case of
invalid input. The application context is recreated before each test
with ddl-auto configured as create-drop to
grant that a test does not impact the result of another and to avoid
problems with PK and unique constraints.