Move authentication business logic from controller to service
The purpuse of this change is to separate the responsabilities, the controller will handle the request and responses and the service will handle the main business logic. This change will also make the core logic of authentication easier to test because it will no longer depend on Spring Web framework.